\docType{package}
\name{somebm}
\alias{somebm}
\alias{somebm-package}
\title{Some functions to generate the time series of Brownian motions.}
\description{
  This package provides some functions to generate the time
  series of Brownian motions, including (regular) Brownian
  motion (\code{\link{bm}}), geometric Brownian motion
  (\code{\link{gbm}}), and fractional Brownian motion
  (\code{\link{fbm}}). They can help users simulate the
  process of one-dimension Brownian motions.
}

